requirements.txt
click==8.1.7
envyaml==1.10.211231
jinja2==3.1.2
pytest-cov==4.1.0
pytest-dotenv==0.5.2
pytest-recording==0.13.0
python-benedict==0.32.1
python-slugify==8.0.1
pyyaml==6.0.1
requests==2.31.0
rich==13.6.0
vcrpy==5.1.0
xlsxwriter==3.1.9
